# apps/core/openapi.yaml
openapi: 3.1.0
info:
  title: Life Engine Core API
  description: >
    JSON REST API for Life Engine Core. Provides data storage (CRUD), full-text
    search, authentication (local-token and OIDC), system health/plugin management,
    quarantine management, Server-Sent Events streaming, and sync conflict resolution.
  version: 0.1.0
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://127.0.0.1:3750
    description: Default local development server
security:
  - bearerAuth: []
...
paths:
...
  /api/events/stream:
    get:
      operationId: eventStream
      summary: SSE event stream
      description: >
        Opens a Server-Sent Events connection. Receives real-time events from
        the internal message bus. Keep-alive pings every 15 seconds.
      tags: [Events]
      parameters:
        - name: collection
          in: query
          schema: { type: string }
          description: Only receive `new_records` events for this collection.
        - name: event_type
          in: query
          schema:
            type: string
            enum: [new_records, sync_complete, plugin_loaded, plugin_error]
          description: Only receive events of this type.
      responses:
        '200':
          description: SSE stream opened.
          content:
            text/event-stream:
              schema:
                type: string
                description: >
                  Standard SSE wire format. Event types: new_records,
                  sync_complete, plugin_loaded, plugin_error.
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ---------------------------------------------------------------------------
  # Conflicts API
  # ---------------------------------------------------------------------------
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Bearer token from `POST /api/auth/token` (local-token mode) or
        `POST /api/auth/login` (OIDC mode).
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: array
              items:
                type: string